refactor: extract erc20 token toggleable type
