+++ src/frontend/src/eth/services/erc20.services.ts
@@ -8,7 +8,8 @@ import { infuraErc20Providers } from '$eth/providers/infura-erc20.providers';
 import { erc20DefaultTokensStore } from '$eth/stores/erc20-default-tokens.store';
 import { erc20UserTokensStore } from '$eth/stores/erc20-user-tokens.store';
 import type { Erc20Contract, Erc20Metadata, Erc20Token } from '$eth/types/erc20';
-import type { Erc20UserToken, Erc20UserTokenState } from '$eth/types/erc20-user-token';
+import type { Erc20UserTokenState } from '$eth/types/erc20-token-toggleable';
+import type { Erc20UserToken } from '$eth/types/erc20-user-token';
 import type { EthereumNetwork } from '$eth/types/network';
 import { mapErc20Token, mapErc20UserToken } from '$eth/utils/erc20.utils';
 import { queryAndUpdate } from '$lib/actors/query.ic';
